+++ Assets/Prefabs/Dialogue UI.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 07cb84f4bf6fcb24a8124e905f2e2071
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
