+++ Assets/Resources/test.csv
@@ -1,0 +1,3 @@
+﻿name,dialog
+2,일어났군요. 아가씨
+2,"이제 곧 미셀스와이트 저택에 도착할거니, \n 아가씨가 가는 곳이 어떤 데인지 이야기해드리는게 좋겠군요."
