+++ Assets/Resources/test.csv.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e4f52264fc40ff646809e35e254f0eb7
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
