+++ Assets/Scripts/Character/NewControll.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9f4bd92f17a163c4fa9da60124de537c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
